pyproject.toml
[project]
name = "biais-llm-rec"
version = "0.1.0"
description = "LLM Recommendation Bias Analysis"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "beautifulsoup4==4.13.3",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "debugpy==1.8.13",
    "decorator==5.2.1",
    "executing==2.2.0",
    "fonttools==4.57.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.0.2",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jupyter-client==8.6.3",
    "jupyter-core==5.7.2",
    "kagglehub==0.3.11",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "numpy>=1.22,<2.1",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.1.0",
    "platformdirs==4.3.7",
    "prompt-toolkit==3.0.50",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.2",
    "pydantic-core==2.33.1",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "redis==5.2.1",
    "requests==2.32.3",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "spotipy==2.25.1",
    "stack-data==0.6.3",
    "tmdbsimple==2.9.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-inspection==0.4.0",
    "typing-extensions==4.13.1",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "thefuzz==0.22.1",
]

[project.optional-dependencies]
vllm = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "airportsdata==20250224",
    "astor==0.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "blake3==1.0.4",
    "cachetools==5.5.2",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "compressed-tensors==0.9.2",
    "cupy-cuda12x>=13.0.0",
    "depyf==0.18.0",
    "dill==0.3.9",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "einops==0.8.1",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "fastrlock==0.8.3",
    "filelock==3.18.0",
    "frozenlist==1.5.0",
    "fsspec==2025.3.2",
    "gguf==0.10.0",
    "httptools==0.6.4",
    "huggingface-hub==0.30.1",
    "importlib-metadata==8.6.1",
    "interegular==0.3.3",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "lark==1.2.2",
    "llguidance==0.7.13",
    "llvmlite==0.43.0",
    "lm-format-enforcer==0.10.11",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mistral-common==1.5.4",
    "mpmath==1.3.0",
    "msgpack==1.1.0",
    "msgspec==0.19.0",
    "multidict==6.3.2",
    "networkx==3.4.2",
    "ninja==1.11.1.4",
    "numba==0.60.0",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "openai==1.70.0",
    "opencv-python-headless==4.11.0.86",
    "outlines==0.1.11",
    "outlines-core==0.1.26",
    "partial-json-parser==0.2.1.1.post5",
    "prometheus-fastapi-instrumentator==7.1.0",
    "prometheus-client==0.21.1",
    "propcache==0.3.1",
    "protobuf==6.30.2",
    "ptyprocess==0.7.0",
    "py-cpuinfo==9.0.0",
    "pycountry==24.6.1",
    "python-dotenv==1.1.0",
    "python-json-logger==3.3.0",
    "python-multipart==0.0.20",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "rich==14.0.0",
    "rich-toolkit==0.14.1",
    "rpds-py==0.24.0",
    "safetensors==0.5.3",
    "scipy==1.15.2",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "starlette==0.46.1",
    "sympy==1.13.1",
    "tiktoken==0.9.0",
    "tmdbsimple==2.9.1",
    "tokenizers==0.21.1",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "transformers==4.50.3",
    "triton==3.2.0",
    "typer==0.15.2",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "vllm==0.8.2",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
    "xformers==0.0.29.post2",
    "xgrammar==0.1.16",
    "yarl==1.18.3",
    "zipp==3.21.0",
]
ollama = [
    "ollama==0.4.7",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["utils"]